+++ resources/js/components/StatusLog/StatusLog.vue
