+++ resources/js/components/StatusLog/StatusLogEvent.vue
